backend/requirements.txt
flask==1.1.1
flask_restful==0.3.7
flask_cors==3.0.8
paho-mqtt==1.5.0
sqlalchemy==1.3.19
psycopg2==2.8.5
